+++ src/app/customForms/customFormConfigModals/customFormWizard/graphEntryPointModal.ts
@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { Component, Input } from "@angular/core";
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { CODAConverter } from "src/app/models/Sheet2RDF";
 import { ConverterConfigStatus } from "src/app/sheet2rdf/s2rdfModals/converterConfig/converterConfigurationComponent";
@@ -8,6 +8,8 @@ import { ConverterConfigStatus } from "src/app/sheet2rdf/s2rdfModals/converterCo
     templateUrl: "./graphEntryPointModal.html",
 })
 export class GraphEntryPointModal {
+
+    @Input() converter: CODAConverter;
 
     converterStatus: ConverterConfigStatus;